+++ core/_build_info.py
@@ -1,3 +1,3 @@
 # Generated at build time — do not edit manually.
-BUILD_TIME = "2026-08-25 19:16"
-BUILD_VERSION = "0.2.0"
+BUILD_TIME = "2026-09-01 17:58"
+BUILD_VERSION = "0.2.1"